Not symlink

+++ calibre_web/rootfs/etc/cont-init.d/00-a_migrate.sh
@@ -6,7 +6,7 @@ set -e
 # Migrate database #
 ####################
 
-if [ -d /homeassistant/addons_config/calibre-web ]; then
+if [ -d /homeassistant/addons_config/calibre-web ] && [ ! -L /homeassistant/addons_config/calibre-web ]; then
     echo "Moving database to new location /config"
     cp -rf /homeassistant/addons_config/calibre-web/* /config/
     rm -r /homeassistant/addons_config/calibre-web
